# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs": {
      "post": {
        "operationId": "JobController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建新任务",
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "operationId": "JobController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Current page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "按状态筛选: PENDING, IN_PROGRESS, COMPLETED, ISSUE, CANCELLED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "按客户ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workerId",
            "required": false,
            "in": "query",
            "description": "按员工ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "开始日期 (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "结束日期 (ISO 8601)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取任务列表（支持筛选和分页）",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}": {
      "get": {
        "operationId": "JobController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取任务详情",
        "tags": [
          "Jobs"
        ]
      },
      "patch": {
        "operationId": "JobController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新任务",
        "tags": [
          "Jobs"
        ]
      },
      "delete": {
        "operationId": "JobController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "取消任务",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/assign": {
      "patch": {
        "operationId": "JobController_assignWorkers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workerIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "workerIds"
                ]
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "指派员工到任务",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/unassign/{workerId}": {
      "patch": {
        "operationId": "JobController_unassignWorker",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "从任务中移除员工",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/check-in": {
      "patch": {
        "operationId": "JobController_checkIn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lat": {
                    "type": "number",
                    "example": 53.3498
                  },
                  "lng": {
                    "type": "number",
                    "example": -6.2603
                  }
                },
                "required": [
                  "lat",
                  "lng"
                ]
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "员工打卡开始工作",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/jobs/{id}/complete": {
      "patch": {
        "operationId": "JobController_completeJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string",
                    "description": "内部备注或现场反馈",
                    "example": "Everything went well, keys returned."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "员工完成任务",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  },
  "info": {
    "title": "CleanOps API",
    "description": "The CleanOps 2026 Management API - 专业清洁业务管理系统",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "scheduledStart": {
            "type": "string",
            "example": "2026-05-01T09:00:00Z"
          },
          "estimatedDuration": {
            "type": "number",
            "example": 120,
            "description": "预计耗时（分钟）"
          },
          "customerId": {
            "type": "string",
            "example": "customer-uuid"
          },
          "workerIds": {
            "example": [
              "worker-uuid-1"
            ],
            "description": "指派的员工ID列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "请带上备用钥匙"
          }
        },
        "required": [
          "scheduledStart",
          "customerId"
        ]
      },
      "UpdateJobDto": {
        "type": "object",
        "properties": {
          "scheduledStart": {
            "type": "string",
            "example": "2026-05-01T09:00:00Z"
          },
          "estimatedDuration": {
            "type": "number",
            "example": 120,
            "description": "预计耗时（分钟）"
          },
          "customerId": {
            "type": "string",
            "example": "customer-uuid"
          },
          "workerIds": {
            "example": [
              "worker-uuid-1"
            ],
            "description": "指派的员工ID列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "请带上备用钥匙"
          }
        }
      },
...
    }
  }
}